+++ docs/fr_FR/changelog.md
@@ -5,6 +5,12 @@
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
 
+# 25/01/2023
+
+- Gestion de la température
+- DEBUG : Gestion de la luminosité
+- DEBUG : Utilisation de minValue et maxValue
+
 # 16/01/2023
 
 - Création du plugin